+++ src/CoolPlace.Tests/Usings.cs
@@ -1,2 +1,1 @@
 global using NUnit.Framework;
-global using CoolPlace.Core;
